+++ apps/infrastructure/src/modules/website/s3_bucket.tf
@@ -1,0 +1,49 @@
+resource "random_integer" "website_bucket_random_integer" {
+  min = 1
+  max = 9999
+}
+
+locals {
+  bucket_name = join("-", ["website-bucket", random_integer.website_bucket_random_integer.result])
+}
+
+resource "aws_s3_bucket" "website" {
+  bucket = local.bucket_name
+}
+
+resource "aws_s3_bucket_public_access_block" "website" {
+  bucket                  = aws_s3_bucket.website.id
+  block_public_acls       = true
+  block_public_policy     = true
+  ignore_public_acls      = true
+  restrict_public_buckets = true
+}
+
+resource "aws_s3_bucket_versioning" "website" {
+  bucket = aws_s3_bucket.website.id
+  versioning_configuration {
+    status = "Suspended"
+  }
+}
+
+resource "aws_s3_bucket_lifecycle_configuration" "website" {
+  # Must have bucket versioning enabled first
+  depends_on = [aws_s3_bucket_versioning.website]
+
+  bucket = aws_s3_bucket.website.id
+
+  rule {
+    id = "Remove noncurrent versions"
+
+    noncurrent_version_expiration {
+      noncurrent_days = 1
+    }
+
+    status = "Enabled"
+  }
+}
+
+resource "aws_s3_bucket_policy" "cloudfront" {
+  bucket = aws_s3_bucket.website.id
+  policy = data.aws_iam_policy_document.website_iam_policy.json
+}