+++ apps/nextjs-website/src/app/[productSlug]/tutorials/[...productTutorialPage]/page.tsx
@@ -20,6 +20,8 @@ import RelatedLinks, {
 } from '@/components/atoms/RelatedLinks/RelatedLinks';
 import { FragmentProvider } from '@/components/organisms/FragmentProvider/FragmentProvider';
 import { Tutorial } from '@/lib/types/tutorialData';
+import ProductBreadcrumbs from '@/components/atoms/ProductBreadcrumbs/ProductBreadcrumbs';
+import { productPageToBreadcrumbs } from '@/helpers/breadcrumbs.helpers';
 
 type Params = {
   productSlug: string;
@@ -111,9 +113,25 @@ const Page = async ({ params }: { params: Params }) => {
   //     product={props.product}
   //     path={props.path}
   //     bannerLinks={props.bannerLinks}
-  //     showBreadcrumbs={false}
   //   >
   //     <FragmentProvider>
+  //       <Box
+  //         sx={{
+  //           maxWidth: '1156px',
+  //           // 80px is the height of the product header
+  //           marginTop: '80px',
+  //           marginX: 'auto',
+  //           paddingTop: 3,
+  //         }}
+  //       >
+  //         <ProductBreadcrumbs
+  //           breadcrumbs={[
+  //             ...productPageToBreadcrumbs(product, props.path, [
+  //               { name: tutorialProps.page.title, path: props.path },
+  //             ]),
+  //           ]}
+  //         />
+  //       </Box>
   //       <Box
   //         sx={{
   //           display: 'flex',
@@ -121,12 +139,11 @@ const Page = async ({ params }: { params: Params }) => {
   //           maxWidth: '1156px',
   //           margin: '0 auto',
   //           paddingBottom: !hasRelatedLinks ? '56px' : 0,
+  //           paddingTop: '56px',
   //         }}
   //       >
   //         <Box
   //           sx={{
-  //             // 78px is the height of the header, 80px is the height of the product header, 24 is the title padding
-  //             paddingTop: '182px',
   //             flexGrow: { lg: 1 },
   //             maxWidth: {
   //               xs: '100%',
@@ -153,14 +170,12 @@ const Page = async ({ params }: { params: Params }) => {
   //               top: 144,
   //             }}
   //           >
-  //             {props.path && (
-  //               <GuideInPageMenu
-  //                 assetsPrefix={props.bodyConfig.assetsPrefix}
-  //                 pagePath={props.path}
-  //                 inPageMenu={props.body}
-  //                 title={translations.productGuidePage.onThisPage}
-  //               />
-  //             )}
+  //             <GuideInPageMenu
+  //               assetsPrefix={props.bodyConfig.assetsPrefix}
+  //               pagePath={props.path}
+  //               inPageMenu={props.body}
+  //               title={translations.productGuidePage.onThisPage}
+  //             />
   //           </Box>
   //         </Box>
   //       </Box>