+++ apps/storybook-app/stories/atoms/CookieBannerScript.xstories.tsx
@@ -1,0 +1,15 @@
+// TODO: Fix error Module not found: Error: Can't resolve '@/config' in 'developer-portal/apps/nextjs-website/src/components/atoms/CookieBannerScript'
+
+// import { Meta, StoryObj } from '@storybook/react';
+// import CookieBannerScript from '../../../nextjs-website/src/components/atoms/CookieBannerScript/CookieBannerScript';
+
+// const meta: Meta<typeof CookieBannerScript> = {
+//   title: 'Atoms/CookieBannerScript',
+//   component: CookieBannerScript,
+// };
+
+// export default meta;
+
+// export const Showcase: StoryObj<typeof CookieBannerScript> = {
+//   args: {},
+// };