+++ apps/strapi-cms/src/api/solution/routes/solution.ts
@@ -1,0 +1,7 @@
+/**
+ * solution router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::solution.solution');