+++ apps/storybook-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # storybook-app
 
+## 0.0.3
+
+### Patch Changes
+
+- f1a2f65: Add website's atoms components' stories
+- fe05e70: Refactor CtaCard to allow an array of tags in input
+- ce380cf: Refactor Feature component to use dark style programmatically
+- c7eb7c7: Add quote component
+
 ## 0.0.2
 
 ### Patch Changes